requirements.txt
django>=4.2,<5.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
psycopg2-binary>=2.9,<3.0
django-cors-headers>=4.3,<5.0
django-ratelimit>=4.1,<5.0
drf-spectacular>=0.27,<1.0
python-decouple>=3.8,<4.0
